requirements.txt
streamlit==1.35.0
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
plotly==5.22.0
torch>=2.1.0
torchvision>=0.16.0